refactor(phase2): Server monitor registry + BotCommandHandler split (#181)
The Server._tick god-method hand-rolled five near-identical
`if X.should_check(): try: results = await X.check(); handle; except: log`
blocks, and Server also owned the whole Telegram bot-command UI. Both are now
decomposed — behavior-preserving.

- Server holds a `list[(monitor, handler)]` built in __init__ (skipping the
  optional ones), and _tick is a single uniform loop. Adding a monitor is one
  list entry — no more editing _tick. Each monitor's result→log/notify logic moved
  into a named `_handle_*_results` coroutine (same messages, same notifications).
- Extracted the inbound bot-command dispatch (`_dispatch_bot_command` + all
  `_cmd_*`) into `notifications/bot_commands.py::BotCommandHandler`, composed by
  Server. `/status` (needs server uptime/channels) stays on Server via a
  `_status_text` callback. Server keeps a thin `_dispatch_bot_command` delegate so
  the existing bot-command tests are unchanged.

server.py ~295→~200 lines. Existing server + bot tests are the regression net
(961 passed); bot_commands.py 100% covered. pyright 0, ruff/docs clean, coverage 81%.
